Revive eglfs' raster window support
A number of consequences of the new rhi-based backingstore
composition were not handled. Most importantly, the fact
that RasterGLSurface is not a thing anymore in practice
causes challenges because we can no longer decide just
based on the surfaceType what a QWindow with OpenGLSurface
would be. (a plain GL window or a GL window with a backing
store?) Also, the backingstore needs to be able to initialize
its backing QRhi by itself, because with eglfs going through
OpenGL is the only way.

Amends 68a4c5da9a080101cccd8a3b2edb1c908da0ca8e

Fixes: QTBUG-102750

+++ src/opengl/qopenglvertexarrayobject.cpp
@@ -274,7 +274,7 @@ void QOpenGLVertexArrayObjectPrivate::destroy()
         vao = 0;
     }
 
-    if (oldContext && oldContextSurface) {
+    if (oldContext && oldContextSurface && oldContextSurface->surfaceHandle()) {
         if (!oldContext->makeCurrent(oldContextSurface))
             qWarning("QOpenGLVertexArrayObject::destroy() failed to restore current context");
     }
